+++ hutool-json/src/main/java/cn/hutool/json/ObjectMapper.java
@@ -113,8 +113,11 @@ public class ObjectMapper {
 			// TODO 过滤器对Bean无效，需补充。
 			mapFromBean(source, jsonObject);
 		} else {
-			// 不支持对象类型转换为JSONObject
-			throw new JSONException("Unsupported type [{}] to JSONObject!", source.getClass());
+			if(false == jsonObject.getConfig().isIgnoreError()){
+				// 不支持对象类型转换为JSONObject
+				throw new JSONException("Unsupported type [{}] to JSONObject!", source.getClass());
+			}
+			// 如果用户选择跳过异常，则跳过此值转换
 		}
 	}