+++ hutool-core/src/main/java/cn/hutool/core/lang/hash/Number128.java
@@ -1,5 +1,7 @@
 package cn.hutool.core.lang.hash;
 
+import java.util.Objects;
+
 /**
  * 128位数字表示，分高位和低位
  *
@@ -86,5 +88,20 @@ public class Number128 extends Number {
 	@Override
 	public double doubleValue() {
 		return longValue();
+	}
+
+	@Override
+	public boolean equals(final Object o) {
+		if (this == o) return true;
+		if (o == null || getClass() != o.getClass()) {
+			return false;
+		}
+		final Number128 number128 = (Number128) o;
+		return lowValue == number128.lowValue && highValue == number128.highValue;
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(lowValue, highValue);
 	}
 }